* Separate Portal from cool modals * move bottom sheet stack to lib directory + add Readme * move Cool Modals to lib dir + Rename cool modals to pan modal (based on Slack PanModal library it's based on)
