+++ ios/fastlane/Appfile
@@ -2,7 +2,7 @@ app_identifier("me.rainbow") # The bundle identifier of your app
 
 itc_team_id("119997837") # App Store Connect Team ID
 
-team_id("L74NQAQB8H") # Developer Portal Team ID
+team_id("L74NQAQB8H") # Developer Index Team ID
 
 # For more information about the Appfile, see:
 #     https://docs.fastlane.tools/advanced/#appfile
