+++ src/components/PortalConsumer.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { LoadingOverlay } from './modal';
 import { useWallets } from '@/hooks';
 import { sheetVerticalOffset } from '@/navigation/effects';
-import { usePortal } from '@/react-native-cool-modals/Portal';
+import { usePortal } from '@/lib/portal';
 
 export default function PortalConsumer() {
   const { isWalletLoading } = useWallets();
