+++ src/lib/bottom-sheet-navigator/README.md
@@ -1,0 +1,4 @@
+# Bottom Sheet Navigator
+
+Custom bottom sheet navigator based on `@gorhom/bottom-sheet` library with Reanimated based modals.
+We are using this navigator on Android for highly interactive modal screens.
