+++ src/lib/bottom-sheet-navigator/views/BottomSheetBackground.tsx
