+++ src/lib/bottom-sheet-navigator/views/BottomSheetNavigatorView.tsx
