+++ src/lib/bottom-sheet-navigator/views/BottomSheetRoute.tsx
@@ -4,7 +4,7 @@ import BottomSheet, {
 } from '@gorhom/bottom-sheet';
 import React, { useCallback, useEffect, useMemo, useRef } from 'react';
 import { Keyboard, View, ViewStyle } from 'react-native';
-import { isKeyboardOpen } from '../../../helpers';
+import { isKeyboardOpen } from '@rainbow-me/helpers';
 import {
   CONTAINER_HEIGHT,
   DEFAULT_BACKDROP_COLOR,
@@ -27,7 +27,7 @@ const BottomSheetRoute = ({
   onDismiss,
   removing = false,
 }: Props) => {
-  //#region extract options
+  // #region extract options
   const {
     enableContentPanningGesture,
     enableHandlePanningGesture,
@@ -38,18 +38,18 @@ const BottomSheetRoute = ({
     height = DEFAULT_HEIGHT,
     offsetY = android ? 20 : 3,
   } = options || {};
-  //#endregion
+  // #endregion
 
-  //#region refs
+  // #region refs
   const ref = useRef<BottomSheet>(null);
 
   const removingRef = useRef(false);
   removingRef.current = removing;
 
   // const
-  //#endregion
+  // #endregion
 
-  //#region styles
+  // #region styles
   const screenContainerStyle: ViewStyle = useMemo(
     () => ({
       bottom: 0,
@@ -66,9 +66,9 @@ const BottomSheetRoute = ({
     }),
     [backdropColor]
   );
-  //#endregion
+  // #endregion
 
-  //#region context methods
+  // #region context methods
   const handleSettingSnapPoints = useCallback(
     (_snapPoints: (string | number)[]) => {
       navigation.setOptions({ snapPoints: _snapPoints });
@@ -105,18 +105,18 @@ const BottomSheetRoute = ({
       handleSettingSnapPoints,
     ]
   );
-  //#endregion
+  // #endregion
 
-  //#region callbacks
+  // #region callbacks
   const handleOnChange = useCallback((index: number) => {
     if (index === -1) {
       onDismiss(routeKey, removingRef.current);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-  //#endregion
+  // #endregion
 
-  //#region effects
+  // #region effects
   useEffect(() => {
     if (removing === true && ref.current) {
       // close keyboard before closing the modal
@@ -129,9 +129,9 @@ const BottomSheetRoute = ({
       }
     }
   }, [removing]);
-  //#endregion
+  // #endregion
 
-  //#region renders
+  // #region renders
   const renderBackdropComponent = useCallback(
     (props: BottomSheetBackdropProps) => (
       <BottomSheetBackdrop
