+++ src/lib/pan-modal/NativeStackView.js
