+++ src/lib/pan-modal/README.md
@@ -1,0 +1,4 @@
+# Pan Modal
+
+Pan modal is an iOS only, native modal component and navigator code based on Slack's PanModal Swift library. It offers
+completely native navigation with bindings to React Navigation.
