+++ src/lib/pan-modal/screens.js
@@ -8,23 +8,23 @@ let AnimatedNativeScreen;
 const ScreensNativeModules = {
   get NativeScreen() {
     NativeScreenValue =
-      NativeScreenValue || requireNativeComponent('RNCMScreen', null);
+      NativeScreenValue ?? requireNativeComponent('RNCMScreen', null);
     return NativeScreenValue;
   },
 
   get NativeScreenStack() {
     NativeScreenStack =
-      NativeScreenStack || requireNativeComponent('RNCMScreenStack', null);
+      NativeScreenStack ?? requireNativeComponent('RNCMScreenStack', null);
     return NativeScreenStack;
   },
 };
 
-function Screen(props, ref) {
+const Screen = (props, ref) => {
   AnimatedNativeScreen =
-    AnimatedNativeScreen ||
+    AnimatedNativeScreen ??
     Animated.createAnimatedComponent(ScreensNativeModules.NativeScreen);
   return <AnimatedNativeScreen {...props} ref={ref} />;
-}
+};
 
 export default {
   Screen: React.forwardRef(Screen),
