+++ src/lib/portal/README.md
@@ -1,0 +1,3 @@
+# Portal
+
+A component that allows displaying stuff over everything else. Natively implemented on iOS and not natively on Android.s
