+++ src/lib/portal/ios/Portal.m
@@ -1,5 +1,5 @@
 //
-//  Portal.m
+//  Index.m
 //  Rainbow
 //
 //  Created by Michał Osadnik on 01/07/2020.
