+++ src/screens/SelectUniqueTokenSheet.tsx
@@ -2,7 +2,7 @@ import { useNavigation, useRoute } from '@react-navigation/core';
 import React, { useCallback, useContext, useEffect } from 'react';
 import RecyclerAssetList2 from '../components/asset-list/RecyclerAssetList2';
 import { SheetHandle } from '../components/sheet';
-import { ModalContext } from '../react-native-cool-modals/NativeStackView';
+import { ModalContext } from '../lib/pan-modal/NativeStackView';
 import { Box } from '@/design-system';
 import { UniqueAsset } from '@/entities';
 import { useWalletSectionsData } from '@/hooks';
